feat: add random funtions to utils package
