+++ packages/utils/src/getRandomNumberInRange.ts
@@ -1,0 +1,2 @@
+export const getRandomNumberInRange = (min: number, max: number) =>
+    Math.floor(Math.random() * (max - min + 1)) + min;
