+++ packages/utils/src/getWeakRandomId.ts
@@ -1,0 +1,8 @@
+export const getWeakRandomId = (length: number) => {
+    let id = '';
+    const list = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+    for (let i = 0; i < length; i++) {
+        id += list.charAt(Math.floor(Math.random() * list.length));
+    }
+    return id;
+};
