+++ packages/utils/src/index.ts
@@ -2,3 +2,5 @@ export * from './createDeferred';
 export * from './arrayPartition';
 export * from './bytesToHumanReadable';
 export * from './mergeObject';
+export * from './getWeakRandomId';
+export * from './getRandomNumberInRange';
